==> step2.ipynb <==
from sentence_transformers import SentenceTransformer
from sklearn.metrics.pairwise import cosine_similarity
import pandas as pd
import numpy as np
from sklearn.cluster import KMeans

# Load the BERT model for sentence embeddings
model = SentenceTransformer('bert-base-nli-mean-tokens')

file_path = 'clean_df.csv'

database = pd.read_csv(file_path)

database['Abstract'] = database['Abstract'].fillna("").astype(str)
database['Title'] = database['Title'].fillna("").astype(str)

database = database[database['Abstract'] != ""]
database = database[database['Title'] != ""]

database['Combined'] = database['Title'] + ". " + database['Abstract']

# Function to get the embeddings for a list of texts
def get_embeddings(texts):
    return np.array(model.encode(texts, show_progress_bar=True))

combined_embeddings = get_embeddings(database['Combined'].tolist())

num_clusters = 10

kmeans = KMeans(n_clusters=num_clusters, random_state=42).fit(combined_embeddings)

cluster_labels = kmeans.labels_

database['Cluster'] = cluster_labels

database.to_csv('clustered_data.csv', index=False)

# Test text

new_text = "Understanding the Workload of Remote Truck Operators with Discrete Event Simulation"
new_text_embedding = get_embeddings([new_text])

# Determine the most suitable cluster
cluster_centers = kmeans.cluster_centers_
similarities = cosine_similarity(new_text_embedding, cluster_centers)
most_similar_cluster = np.argmax(similarities)

print(f"The most close cluster is: {most_similar_cluster}")

# Calculate similarity to all embeddings in the database
similarities_to_all = cosine_similarity(new_text_embedding, combined_embeddings).flatten()

# Get the five indexes with the highest similarity
top_five_indices = np.argsort(similarities_to_all)[-5:]

top_five_titles = database.iloc[top_five_indices]['Title']
top_five_clusters = database.iloc[top_five_indices]['Cluster']

print(f"The titles of the most similar contents to the given text, their titles, and clusters are:")
for title, cluster in zip(top_five_titles, top_five_clusters):
    print(f"- {title} (Cluster {cluster})")

from sklearn.metrics import silhouette_score, calinski_harabasz_score, davies_bouldin_score

# Silhouette Score
silhouette_avg = silhouette_score(combined_embeddings, cluster_labels)
print("Silhouette Score: ", silhouette_avg)

# Calinski-Harabasz Index
calinski_harabasz = calinski_harabasz_score(combined_embeddings, cluster_labels)
print("Calinski-Harabasz Index: ", calinski_harabasz)

# Davies-Bouldin Index
davies_bouldin = davies_bouldin_score(combined_embeddings, cluster_labels)
print("Davies-Bouldin Index: ", davies_bouldin)

database
